Added validations and image for product

+++ EcomAppSearchableDesign/Controllers/OrderController.cs
@@ -14,30 +14,52 @@ public class OrderController(IOrderService orderService) : ControllerBase
     [HttpPost("create-order")]
     public async Task<HttpResponses<OrderResponseDto>> CreateOrder([FromBody] OrderCreateDto dto, CancellationToken cancellationToken)
     {
+
+        if (!ModelState.IsValid)
+        {
+            return ModelState.ToErrorResponse<OrderResponseDto>();
+        }
         return await orderService.CreateOrderAsync(dto, cancellationToken);
     }
 
     [HttpGet("get-order/{id:int}")]
     public async Task<HttpResponses<OrderResponseDto>> GetOrder(int id, CancellationToken cancellationToken)
     {
+
+        if (!ModelState.IsValid)
+        {
+            return ModelState.ToErrorResponse<OrderResponseDto>();
+        }
         return await orderService.GetOrderByIdAsync(id, cancellationToken);
     }
 
     [HttpPost("get-orders")]
     public async Task<HttpResponses<PagedResult<OrderResponseDto>>> GetOrders([FromBody] GetOrdersDto dto, CancellationToken cancellationToken)
     {
+        if (!ModelState.IsValid)
+        {
+            return ModelState.ToErrorResponse<PagedResult<OrderResponseDto>>();
+        }
         return await orderService.GetOrdersAsync(dto, cancellationToken);
     }
 
     [HttpPut("update-order/{id:int}")]
     public async Task<HttpResponses<OrderResponseDto>> UpdateOrder(int id, [FromBody] OrderUpdateDto dto, CancellationToken cancellationToken)
     {
+        if (!ModelState.IsValid)
+        {
+            return ModelState.ToErrorResponse<OrderResponseDto>();
+        }
         return await orderService.UpdateOrderAsync(id, dto, cancellationToken);
     }
 
     [HttpDelete("delete-order/{id:int}")]
     public async Task<HttpResponses<string>> DeleteOrder(int id, CancellationToken cancellationToken)
     {
+        if (id==0)
+        {
+            return ModelState.ToErrorResponse<string>();
+        }
         return await orderService.DeleteOrderAsync(id, cancellationToken);
     }
